pyproject.toml
[project]
name = "finch"
version = "1.0.0a1"
description = "Open source and cross-platform GUI client for Amazon S3 and compatible storage platforms"
authors = [
    {name = "Furkan Kalkan", email = "furkankalkan@mantis.com.tr"},
]
dependencies = [
    "PySide6>=6.7",
    "boto3==1.35.5",
    "keyring==25.3.0",
    "python-slugify==8.0.4",
]
readme = "README.md"
license = {text = "MIT"}
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: Microsoft",
    "Operating System :: POSIX",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Topic :: Internet",
    "Topic :: Software Development",
    "Topic :: Utilities",
]


[project.urls]
Homepage = "https://github.com/mantis-software-company/finch"

[project.scripts]
finch = "finch.__main__:main"

[tool.setuptools.packages]
find = {}

[tool.setuptools.package-data]
"finch" = ["img/*.png", "img/*.svg", "img/*.qss"]


[tool.cxfreeze]
executables = [
    {script = "finch.py", base = "gui", target_name="Finch S3 Client", copyright="Copyright (C) 2024 Furkan Kalkan", icon = "icon.ico", shortcut_name="Finch S3 Client", shortcut_dir='StartMenuFolder'}
]


[tool.cxfreeze.build_exe]
include_msvcr = true
include_files = [
    ["finch/img/theme.qss", "lib/finch/img/theme.qss"],
]
excludes = [
    "tkinter",
    "unittest",
    "pydoc",
    ]

zip_include_packages = ["encodings", "PySide6", "shiboken6"]

includes = [
    'keyring.backends.SecretService',
    'keyring.backends.libsecret',
    'keyring.backends.chainer',
    'keyring.backends.Windows',
    'keyring.backends.kwallet',
    'keyring.backends.macOS',
]

[tool.cxfreeze.bdist_mac]
iconfile = "icon.png"
bundle_name = "Finch S3 Client"
codesign_identity = "Developer ID Application: Furkan Kalkan (3YR4725H5N)"
codesign_options = "runtime"
codesign_deep = true
codesign_timestamp = true
codesign_verify = true


[tool.cxfreeze.bdist_dmg]
applications_shortcut = true
volume_label = "Finch S3 Client"
background = "builtin-arrow"

[tool.cxfreeze.bdist_msi]
install_icon = "icon.ico"
